python/requirements.txt
Flask
langchain
langchain-community
sentence-transformers
faiss-cpu
openai
python-dotenv
flask_cors